move database upgrade files to the new databaseUpgrades folder
