+++ acquamanCommon.pri
@@ -151,6 +151,12 @@ HEADERS += \
 	source/ui/dataman/AM2DScanView.h \
 	source/ui/AMDatamanStartupSplashScreen.h \
 	source/dataman/database/AMDbUpgrade.h \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt1.h \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt2.h \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt3.h \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt4.h \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt5.h \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt6.h \
 	source/dataman/datasource/AMDataSourceImageDatawDefault.h \
 	source/util/AMDataSourcePlotSettings.h \
 	source/util/AMSelectablePeriodicTable.h \
@@ -212,8 +218,6 @@ HEADERS += \
 	source/dataman/info/AMOldDetectorInfo.h \
 	source/beamline/AMOldDetector.h \
 	source/dataman/info/AMOldDetectorInfoSet.h \
-	source/dataman/AMDbUpgrade1Pt1.h \
-	source/dataman/AMDbUpgrade1Pt2.h \
 	source/ui/beamline/AMOldDetectorView.h \
 	source/ui/beamline/AMOldDetectorViewSupport.h \
 	source/ui/beamline/AMDetectorView.h \
@@ -249,7 +253,6 @@ HEADERS += \
 	source/beamline/AMMotorGroup.h \
 	source/ui/AMMotorGroupView.h \
 	source/util/AMPointerTree.h \
-	source/dataman/AMDbUpgrade1Pt4.h \
 	source/dataman/AMSamplePre2013.h \
 	source/dataman/AMSamplePlatePre2013.h \
 	source/ui/dataman/AMSamplePlatePre2013View.h \
@@ -331,7 +334,6 @@ HEADERS += \
 	source/analysis/AM1DDarkCurrentCorrectionABEditor.h \
 	source/beamline/AMCurrentAmplifier.h \
 	source/beamline/AMAdvancedControlDetectorEmulator.h \
-	source/dataman/AMDbUpgrade1Pt3.h \
 	source/acquaman/AMScanActionControllerBasicFileWriter.h \
 	source/acquaman/AMTimedScanActionController.h \
 	source/acquaman/AMTimedRegionScanConfiguration.h \
@@ -378,7 +380,6 @@ HEADERS += \
 	source/ui/util/AMAboutDialog.h \
 	source/beamline/AM1DControlDetectorEmulator.h \
 	source/ui/actions3/AMActionRunnerCurrentViewBase.h \
-	source/dataman/AMDbUpgrade1Pt5.h \
 	source/beamline/AMBeamlineControlAPI.h \
 	source/beamline/AMBeamlineControlSetAPI.h \
 	source/beamline/AMBeamlineDetectorAPI.h \
@@ -458,7 +459,6 @@ HEADERS += \
 	source/util/AMTimer.h \
 	$$PWD/source/acquaman/AMXRFScanConfiguration.h \
     $$PWD/source/acquaman/AMXRFScanController.h \
-	$$PWD/source/dataman/AMDbUpgrade1Pt6.h \
 	source/ui/beamline/AMControlView.h \
     source/beamline/AMSlit.h \
     source/beamline/AMSlitGap.h \
@@ -587,6 +587,12 @@ SOURCES += \
 	source/ui/dataman/AM2DScanView.cpp \
 	source/ui/AMDatamanStartupSplashScreen.cpp \
 	source/dataman/database/AMDbUpgrade.cpp \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt1.cpp \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt2.cpp \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt3.cpp \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt4.cpp \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt5.cpp \
+	source/dataman/databaseUpgrades/AMDbUpgrade1Pt6.cpp \
 	source/dataman/datasource/AMDataSourceImageDatawDefault.cpp \
 	source/util/AMSelectablePeriodicTable.cpp \
 	source/ui/util/AMSelectablePeriodicTableView.cpp \
@@ -646,8 +652,6 @@ SOURCES += \
 	source/dataman/info/AMOldDetectorInfo.cpp \
 	source/beamline/AMOldDetector.cpp \
 	source/dataman/info/AMOldDetectorInfoSet.cpp \
-	source/dataman/AMDbUpgrade1Pt1.cpp \
-	source/dataman/AMDbUpgrade1Pt2.cpp \
 	source/ui/beamline/AMOldDetectorViewSupport.cpp \
 	source/ui/beamline/AMOldDetectorView.cpp \
 	source/ui/beamline/AMDetectorView.cpp \
@@ -683,7 +687,6 @@ SOURCES += \
 	source/beamline/AMMotorGroup.cpp \
 	source/ui/AMMotorGroupView.cpp \
 	source/util/AMPointerTree.cpp \
-	source/dataman/AMDbUpgrade1Pt4.cpp \
 	source/dataman/AMSamplePre2013.cpp \
 	source/dataman/AMSamplePlatePre2013.cpp \
 	source/ui/dataman/AMSamplePlatePre2013View.cpp \
@@ -763,7 +766,6 @@ SOURCES += \
 	source/analysis/AM1DDarkCurrentCorrectionABEditor.cpp \
 	source/beamline/AMCurrentAmplifier.cpp \
 	source/beamline/AMAdvancedControlDetectorEmulator.cpp \
-	source/dataman/AMDbUpgrade1Pt3.cpp \
 	source/acquaman/AMScanActionControllerBasicFileWriter.cpp \
 	source/acquaman/AMTimedScanActionController.cpp \
 	source/acquaman/AMTimedRegionScanConfiguration.cpp \
@@ -810,7 +812,6 @@ SOURCES += \
 	source/ui/util/AMAboutDialog.cpp \
 	source/beamline/AM1DControlDetectorEmulator.cpp \
 	source/ui/actions3/AMActionRunnerCurrentViewBase.cpp \
-	source/dataman/AMDbUpgrade1Pt5.cpp \
 	source/beamline/AMBeamlineControlAPI.cpp \
 	source/beamline/AMBeamlineControlSetAPI.cpp \
 	source/beamline/AMBeamlineDetectorAPI.cpp \
@@ -890,8 +891,7 @@ SOURCES += \
     source/util/AMStorageInfo.cpp \
 	source/util/AMTimer.cpp \
 	$$PWD/source/acquaman/AMXRFScanConfiguration.cpp \
-    $$PWD/source/acquaman/AMXRFScanController.cpp \
-    $$PWD/source/dataman/AMDbUpgrade1Pt6.cpp \
+	$$PWD/source/acquaman/AMXRFScanController.cpp \
 	source/ui/beamline/AMControlView.cpp \
     source/beamline/AMSlit.cpp \
     source/beamline/AMSlitGap.cpp \
